+++ lib/flutter_pay.dart
@@ -1,10 +1,10 @@
 library flutter_pay;
 
-import 'dart:async';
+import 'dart:io';
 import 'package:flutter/services.dart';
 
-part 'models/flutter_pay.dart';
-part 'models/flutter_pay_error.dart';
-part 'models/payment_environment.dart';
-part 'models/payment_item.dart';
-part 'models/payment_network.dart';
+part 'src/flutter_pay.dart';
+part 'src/flutter_pay_error.dart';
+part 'src/payment_environment.dart';
+part 'src/payment_item.dart';
+part 'src/payment_network.dart';
