+++ lib/src/flutter_pay_error.dart
