+++ lib/src/payment_item.dart
